+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog for the Cortex terraform provider.
 
 ## Unreleased
 
+* Add `wiz` attribute support for Catalog Entities
 * Add `firehydrant` attribute support for Catalog Entities
 * Add the ability to ignore metadata on a Catalog Entity, allowing it to be managed via API instead